refactor: reorganize reusable components into common folder structure
