+++ src/components/TodoListRow.tsx
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector } from 'react-redux';
 import { HiPencil, HiTrash } from 'react-icons/hi2';
-import Checkbox from './Checkbox';
+import Checkbox from '@common/Checkbox';
 import { deleteTodoAsync, toggleTodoAsync } from '@src/redux/todosSlice';
 import { showEditModal } from '@src/redux/modalSlice';
 import formatDateString from '@utils/formatDateString';
