+++ src/components/common/LoadingSpinner/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LoadingSpinner';
